MB-54516: Modify a replica collection over DCP (in-memory only)
Commit enables DCP to produce a collection modify system event
over DCP - for in-memory streaming only. With this commit a
history change made to the active collection will generate an
event which DCP will transmit to clients that have opted-into
flatbuffers_system_events

Support for flushing the event (and backfill testing) comes in
a later change.

Change-Id: I7ac0955c6242df63ee23c63c3d2ae34eec8a1d4d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183250
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>