MB-54645: Enable operation tracing by default
+Background+

As seen in recent CBSEs, not all SDK clients enable HELO::Tracing on
their mcbp connections - either because they don't enable it by
default (e.g. Query prior to MB-64530 / 7.5.0), or because the
application has explicitly disabled it.

If Tracing is not enabled then KV-Engine does not generate trace spans
for the low-level operation phases if they are not already timed - and
as such it can be confusing to try to debug why these "Slow
operations" were slow.

One example of this is the bg.wait / bg.load spans which can be used
to identify operations which went to disk, and how much time was spent
waiting on disk - these are currently only recorded (and reported in
Slow op message) if the client enabled tracing.

+Change+

This patch unconditionally collects tracing information by setting
memcached.json always_collect_trace_info to true. This guarantees we
always have detailed trace information (for the (small) number of
clients which don't already enable it).

Note that this setting differs from an explicit HELLO::Tracing in that
it doesn't automatically include the server duration in the response,
so there's no change in behavour in the mcbp wire protocol - i.e. it's
totally backwards compatible for clients.

Change-Id: Ie9135907121fdd86c53622b03730150ca1ac2435
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183124
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>