MB-54516: Support flush of a ModifyCollection system event
Update the schema in kvstore.fbs to include a new bool field for storing
the collections history setting.

The KVStore derived VB::Manifest now captures the history setting from
CreateCollection events.

With that, check for the flush of a ModifyCollection event and use the
settings from that event to make any adjustments to the KVStore
collection data.

VB::Manifest now reloads the correct state from warmup and flush
support allows testing of Modify from backfill.

A limitation of this commit is that drop collection does not yet
support the ModifyCollection event (it is not dropped when it
should be).

Change-Id: I523a4a0ea2c2b3de70d7f6cc621936495a6f8cb1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183251
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>