kv_slow_ops_2_gtrace: Move tests after production code
Recent additions to this file have meant the tests were in the middle
of the main code.

Move them to the end, just before main() - which has to be at the end
for the automatic test registration to work.

Also remove a leftover 'print(obj) in one of the unit tests.

Change-Id: I7bb8f41674de7a422ff8bd18d537738e1ccadae9
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183223
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>