kv_slow_ops_2_gtrace: Prefer opaque from packet field
clientID /should/ be a tuple of "connectionID/opaque", but not all
clients obey this convention. As such, if the trace has the
packet.opaque field (newer server versions added it), use that
instead.

(Also include worker_tid in the output as I spotted that was missed
out when added to the server output).

Change-Id: Iebf33caf296878c937ef3415c4e66b5ead4aa98c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183125
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>