MB-54700: Reduce the size of StoredValues by 4 bytes
By changing the datatype to be a 3-bit long bit field and combining it
with the 5-bits of flags that we had, and ignoring the 4 bytes of
resulting padding when allocating StoredValues.

It is difficult to precisely quantify the benefit of this change -- we
use jemalloc in our release builds and it uses bins of size 64, 80, 96,
112, 128... Some of the allocations will go in the lower bin size.

For example, if the document key is 23-26 bytes long, these allocations
will be in the 80 byte bin, whereas before this change they would have
ended up in the 96 byte bin, giving a saving of 16 bytes per StoredValue
with that key size.

Note that OrderedStoredValue size is unaffected (and the compiler adds
the appropriate padding for the base class object).

Change-Id: I6cc52ff5a2db4dd253a6d13eab2bb17e6bf035b9
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183284
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>