Refactor: migrate to std::filesystem
Use std::filesystem::remove[_all] instead of cb::io::rmrf
it does not throw exception if the target doesn't exist so
we don't need to catch the exception and check for ENOENT
and possibly throw the exception again.

Use std::filesystem::create_directories instead of
cb::io::mkdirp as it does the same thing.

Change-Id: Ibf3cc5636f3bb1938862ee324e69259b9c0c2c3e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183341
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>