Use the logger directly from cbsasl
Previously memcached would set a log function in cbsasl and in cbsasl
we would (unconditionally) format the log output and then call the
callback to log the content.

This patch use the logger directly from cbsasl with the log macros
which would bypass formatting the content if the log entry is supposed
to be dropped.

Lower the log messages upon invalid input provided from the client
to DEBUG instead of ERROR (as it only affects clients which send
bad data to the server and we don't want such clients to be
able to flood the server log).

Change-Id: I1e0ebca2379ccbc41b26c0597cd1e24586fefffb
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183371
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>