MB-52823: Disable test of TLS1[.1] for OpenSSL 3.0
The unit tests for TLS 1 and TLS 1.1 fails as the code currently
stands when running under OpenSSL 3.0. To ease the migration to
OpenSSL 3.0 disable the tests for now if OpenSSL 3.0 is being
used.

Once the server is building with OpenSSL 3.0 we should consider
if we want to _remove_ support for TLS 1 and 1.1 (as it is old
and should be avoided) or if we want to update the code so that
they work.

Change-Id: I34bb5b24fdc92a4d79acdc2286c2403c8a616172
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183423
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>