DCP docs: Add 'design discussion' page
Add a new 'DCP Design Discussion' page, to cover additional details on
why DCP works in certain ways, and gives examples of how it can break
if certain rules are not followed.

The page initially has details on why snapshot start/end must be
passed correctly when resuming a stream.

Change-Id: Ie5a48ecd8a9d1d79444c6a02bbf61085e68a0565
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183611
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>