MB-53945: [linux] Add support for TCP_USER_TIMEOUT
This only applies to clients connecting to the user interface(s)
(11210 etc) and not the system interface(s) (11209 etc).

TCP_USER_TIMEOUT solves the "problem" where alive clients
(both ends of the socket is alive and TCP Keepalive probes
gets sent and replies received) but for some reason the
client don't read data off the socket causing the data
to be buffered inside the kernel TCP buffer on the server
without any progress.

The value must be set in memcached.json through the key:

    "tcp_user_timeout" : <number of seconds>

Setting the value to 0 == disabled

Change-Id: Ie59353d00dbe9be07485d5aa3f04b7b242cb2c4a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183701
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>