MB-54516: Replace ModifyCollection with SeqnoAdvance
For connections which have enabled collections, but not FlatBuffers
the ModifyCollection event is skipped, SeqnoAdvance can replace the
skipped seqno to ensure snapshots are closed.

Note that KV-KV replication this does not apply. A ModifyCollection
only occurs once all nodes are 7.2+ (ns_server cluster_compat). A
SeqnoAdvance is not required for replication streams (note KV would
reject it).

Change-Id: I97f9f1c16a8fb084c0dbed28f46b4e008d114982
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/184077
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>