MB-54516: Modify history out of sync with many vbuckets
The initial addition of history mistakenly added the flag to the
"shared metadata". When >1 active vbucket exists, the first one
to be processed updated collection.history, but sets the shared
metadata and then generates a Modify. The second vbucket sees
no difference in history and does not generate a Modify.

The replicas get out of sync, with vb0 correctly deduplicating
on active and replica, but vb1 only deduplicating on the active.

Change-Id: If653494f31e44c2b9fb44f45d07bd814f639a46a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/184179
Well-Formed: Restriction Checker
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>