Refactor: ItemPager::createVBucketFilter returns nullopt for no match
...because an empty VBucketFilter is actually matches all vBuckets.

This refactor also exposed an undersirable behaviour which was depended
on by one of the tests, and has been relied upon since the test has
existed -- if there are no vBuckets in an evictable state, item pager
used to attempt to evict items from all vBuckets, because the empty
VBucketFilter that it had matched everything. The dead vBucket in that
test was ignored as expected, because the pauseResumeVisit code checks
whether the bucket is dead, but importantly, the item pager kept
rescheduling.

This is no longer the case. Added a workaround to trigger the item pager
as the test expected.

Change-Id: Ife0ff04f433e55701f9ad876044813fbd4f41767
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/184251
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>