MB-54958: Move the protocol bits of Observe to the core
Separate out the bits for the protocol encoding from ep-engine
and perform that in the core.

Given that the core supports a state machinery for the command
we don't need to redo all keys every time we have to block for
a key; and may continue to the next key.

Change-Id: I2339309b5e4f741f0bef54eb3b72487b5204f779
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/184459
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>