MB-64344: Move the protocol bits of Observe to the core
This is a backport of 6bdc02f from MB-54958 to 7.2.x, with the
additional fix in MB-60726 (8fffa06).

Separate out the bits for the protocol encoding from ep-engine
and perform that in the core.

Given that the core supports a state machinery for the command
we don't need to redo all keys every time we have to block for
a key; and may continue to the next key.

Change-Id: I2339309b5e4f741f0bef54eb3b72487b5204f779
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/219887
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>