MB-54729: Add historyStartSeqno to ScanContext
Add a new member, historyStartSeqno to ScanContext for later use in
DCP backfill when "change_stream=true".

This will always be 0 for a KVStore which reports:

    StorageProperties::HistoryRetentionAvailable::No

For a KVStore which reports:

    StorageProperties::HistoryRetentionAvailable::Yes

This will be set to:

    0: If no history is configured (retention_bytes=0)
    >0: When history is available, this value represents the lowest
        seqno at which a continuous change stream could be made
        available.

This commit is not dependent upon Magma's changes to add history, but
adds the ability for MockMagmaKVStore to "inject" a history start.

Change-Id: I66db1d84af54f4a999f260d30b356a9a149ddf31
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/184674
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>