Move withLabels to base StatCollector
withLabels allows construction of a new collector which will add the
provided labels to every subsequent addStat call:

 StatCollector c = ...;
 auto labelled = c.withLabels({{"foo", "bar"}});
 labelled.addStat(Key::something, value);
 labelled.addStat(Key::something_else, otherValue);

However, this was previously only defined on LabelledStatCollector
itself, so was only usable on Bucket/Scope/CollectionStatCollectors.

That's an unnecessary restriction; indeed, prior to
https://review.couchbase.org/c/kv_engine/+/137974
withLabels was available on StatCollector.

Restore StatCollector::withLabels.

Change-Id: Ifea59a9617f361621e2f5eb07141a5d3dd025ca4
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/184525
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>