MB-54729: Add snapshot flag "may contain duplicates" for CDC backfill
Adds: 0x20 May Contain Duplicates

The flag "May Contain Duplicates" is added so that DCP can choose to
produce snapshots that change the "unique-key" definition. The usage
of this flag will be for disk source snapshots where the cursor is
within the window of history. A single snapshot will be produced and
it can store many versions of a key.

Change-Id: I70c80f799a469247e12873bf475cec00875c1f75
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/184666
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>