Revert "MB-54876: Add support for magma dynamic block sizes"
This reverts commit e019d40fa8cf76b81216a263f4b2bf09082eeb3e.

Reason for revert: This breaks the elixir build 
(https://server.jenkins.couchbase.com/job/couchbase-server-unix/394577/console) as while kv_engine is using the master branch there, magma is not and instead is using the elixir branch, which doesn't have the corresponding Magma patch (https://review.couchbase.org/c/magma/+/184439) included in it yet.

If we want this feature for 7.2.0, then Magma needs to merge patch 184439 to magma/elixir branch before re-submitting this. 

Change-Id: I294b302c30cd5e18730a6349e5f41f5c30e91eb7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/183694
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>