MB-54668: Remove magma_value_separation_size config
Note: This is a forward port of fa6913752 that is "merged" out of
order with the rest of neo merging so that all of KV's dependent
magma changes can be merged to magma/master.

Magma stored small values under the magma_value_separation_size
in the keyIndex and larger values in the seqIndex log. This
resulted in addiotnal complexity with negligible benefit since
the seqno still have to be stored in the seqIndex.
Also this needs to be disabled for Magma to retain history
since the seqIndex is where Magma will pause compaction.

Removing magma_value_separation_size and the MinValueSize
config such that all values are stored in the seqIndex. Magma
has retained the CanSplitValueFn to allow for testing of the
read path for existing Magma instances.

Change-Id: I56f6a1bdaf0d7f5d7309c63257e6883b3b2d3805
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/187198
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>