MB-54668: Remove magma_value_separation_size config
Magma stored small values under the magma_value_separation_size
in the keyIndex and larger values in the seqIndex log. This
resulted in addiotnal complexity with negligible benefit since
the seqno still have to be stored in the seqIndex.
Also this needs to be disabled for Magma to retain history
since the seqIndex is where Magma will pause compaction.

Removing magma_value_separation_size and the MinValueSize
config such that all values are stored in the seqIndex. Magma
has retained the CanSplitValueFn to allow for testing of the
read path for existing Magma instances.

Change-Id: I5a05b74454fdc5d6167f36644e69151cfc775028
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/184790
Well-Formed: Restriction Checker
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>