MB-54850: Skip magma key-lookup optimization if Disk snap is Historical
As an optimization for improving the performance at flush, we tell
magma whether a replica is receiving the initial disk snapshot, ie the
vbucket is empty. In that case magma can skip the key lookups required
for maintaining the item-count, as any mutation received is an Insert
by logic.

That isn't the case anymore for when the initial disk snapshot is
Historical, as the snapshot might contain duplicates.

Change-Id: I453339fd9dcbe9cbd4dcd8f4c54a18cd8543d5ca
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/184967
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>