MB-51979: Utilise FlatBuffers SystemEvents to clean-up Metered handling
When Metered was added as new collection metadata two paths existed for
ensuring replicas saw the correct metering state.

1) Allow the replica and active to differ, "fixing" the replica state
when a switch to active occurs.

2) Fix all collection creation paths so that they could replicate the
collection's metered state.

At the time 1 was implemented as it could borrow some code from scope
data limits and the overall cost/benefit was better then 2.

However the collection history setting added in neo forced our hand to
implement the infrastructure to make 2 possible - FlatBuffers system
events allow Create collection to easily expand and include new fields.

This commit now exists to simplify the metered management and make way
for a merge of neo "ModifyCollection" code which is required for
the history flag. With this commit, metered is now added to the
FlatBuffers event schema and replicated between KV nodes, the replica
now gets the correct state and has no need to check during the switch
to active.

Change-Id: Id14f0bdfc2c4be6bf442182f1ab371599df90224
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/184798
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>