MB-54516: Update collection history configuration
Primary change is that all history enablement is only via the
collection entry, removing any bucket/scope entry overrides.

Also changed in this commit is the disablement of history on the
default collection. Required to avoid having to support modify
collection with getAllVBSeqnos and a legacy DCP client.

Change-Id: Ie5c9c425bad50f76f2b874b2eb92491a3cc43f3f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/184804
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>