MB-54729: CDC stream backfill with a dropped collection
Primarily add a test that includes a dropped collection in the
scanned history range. This highlighted a problem in the collections
ScanContext as with history enabled the scanAllVersions method will
return both the create and drop of a collection as separate system
events. The isLogicallyDeleted code now needs to be able to distinguish
the create and drop collection events so that a backfill can exclude
the create event if the collection is dropped.

Change-Id: Icd78e81e9106d7c7014e9e7caecf89a948960d24
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/185020
Well-Formed: Restriction Checker
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>