MB-53053: Change ownership model for CrossBucketVisitorAdapter
Make the CrossBucketVistorAdapter enable_shared_from_this and have the
spawned visitor tasks take a strong reference to the adapter, while the
adapter gets a weak reference to the tasks.

This fixes a use-after-free which could happen due to a race condition
in the QuotaSharingVisitorAdapter when all of the following are true:
- The engine is shutting down, so we ignore the task in scheduleNext
- CrossBucketVisitorAdapter::completed gets set to true
- The shutting down engine schedules the task to run because of
  blocksShutdown, but the CrossBucketVisitorAdapter was destroyed after
  being removed from releaseCompletedAdapters.

To fix this, change the ownership model of this object. Make the
adapter have shared ownership between its spawned visitor tasks. By
doing this, we no longer need the releaseCompletedAdapters.

Once all spawned tasks have been destroyed, the adapter itself is
destroyed.

Change-Id: I505f5ff2845703f83f5739335764467a8f86e208
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/185120
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>