Fix inaccurate memory usage checks in paging tests
After making some changes to the QSPager, some of these tests started to
fail, complaining in the assertion after the populateVbsUntil that the
memory usage is now 4 bytes below the high_wat (when the predicate it
checks just checked that the memory usage is strictly above the
high_wat).

The changes I made reduced the size of the QSPager task. After some fun
debugging I realised that until now, we were relying on some planetary
alignment which caused us to write enough items such that the mentioned
heap allocation by the temporary vbids vector was a non-issue.

Change-Id: Ia9f5f238763a4cd208412a222b6cbfbf5412e876
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/185099
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>