MB-54553: Propagate CheckpointHistorical to magma::saveDocs()
Passed when committing the flush-batch.
Primary usage is to inform magma in the case where the historical
data stream has been interrupted.

* Note *
Magma needs that the flag is always passed to the storage.
7.2 active nodes always create checkpoints with Historical::Yes.
While 7.2 replicas set the Historical flag in checkpoints depending on
what is received in snapshot Markers.

Change-Id: Ia1a3ff07130a77e2cc919334adee1876c0092c9e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/185333
Well-Formed: Restriction Checker
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>