MB-52682: Allow an Item's freqCounterValue to be left unspecified
We want to be able to use the per-vBucket MFU histograms to choose the
starting MFU values for items stored in the HashTable.

To do that, we need to be able to distinguish between Items for which we
have explicitly set this value (like for items received over DCP) vs
items coming in from the front-end.

Make the freqCounterValue accessors use std::optional to distingush
between those two cases.

Change-Id: I1fc59be125730ca3121b3d3587ba7881762e4db0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/185221
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>