Run the cluster_test serially
After some debugging on Windows, I've established that the following
error messages or some combination can be seen when we run out of
memory:

- [warn] event_base_new_with_config: Unable to make base notifiable.
- unknown file: error: C++ exception with description "Failed to create
  pipe for AtomicNotificationQueue9: bad file descriptor"
- [warn] evsig_init_: socketpair: Only one usage of each socket address
  (protocol/network address/port) is normally permitted.
- unknown file: error: SEH exception with code 0xc0000005

In an effort to address those, make the cluster_test run serially. Why
target the cluster_test specifically -- it spins up 4 memcached
instances, so it is one of the heavier tests.

Change-Id: If36ba3d21ece9fcdba3d0aecc5ecb4f711507ebd
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/185270
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>