MB-54553: Handle NonHistory->History Memory snap transition at replica
When a History stream is interrupted on a 7.2 node, then we need to
inform magma about that. That is implemented by passing the History
flag within the flush-batch to magma.
That implies that we can never merge checkpoints with different History
configuration in the flush-batch.
While that already applies to Disk checkpoints, this patch introduces
the same logic for Memory checkpoints.

Note that the use case here is when a replica vbucket sees a NonHistory
to History jump in the inbound stream. At the time of writing that is
expected to happen only in the pre7.2->7.2 online upgrade scenario.

Change-Id: Ie41605ce0126624ee66feeafcee8207bd424bca0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/185307
Well-Formed: Restriction Checker
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>