MB-54729: Fix backfill that begins within the history window
The initial code for MB-54729 was flawed and did not correctly support
a change stream backfill when the start lands inside the history
window.

The sequence numbers which are recorded and then used in the history
scan were incorrect, meaning the history scan was breaking monotonic
expectations of the DCP producer.

A new test covers the broken case.

Change-Id: I0d3c7cef71ad15c390401b018270f5daf7c3c9c2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/185544
Well-Formed: Restriction Checker
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>