MB-54729: Enable history scan for CDC backfill
Replace the todo markers with code that now utilises the magma history
API - this now means scanAllVersions for example is hooked into the
magma history scanning API.

Add new tests that validate multiple versions can be stored and
returned.

Also required are changes to unit tests to respect new expectation
checks that occur in magma - primarily that flushing writes ordered
batches - this is only a problem for tests which bypass the flusher
and call KVStore directly.

Change-Id: Ife0bfaf6b0fb8dade115c879b96733c2b70b1e6f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/185933
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>