Add comment about motivation for flow-control.md
Extend the documentation of flow control for an additional reason
why you want flow control and not rely on TCP flow control (and
one of the reasons we did it).

If one relies on TCP flow control and the buffer is "full" (you're
blocked writing) you have no idea if the _application_ on the other
side is hung or working as "expected" from your point of view. All
we know is that data towards the other machine is stuck. By using
our own flow control we can verify that the entire network stack
on our side, the network path to the other machine and up the
other stack all the way into the application is working if we
haven't heard from the other side for a period. That would be
"impossible" if we're stuck writing 20MB into the kernel buffer.

"back in the days" we had some interesting network issues when
testing on EC2 where the application on node A was blocked
writing data, and the same thing on the other side.. but the
network in between was not "working". Thats most likely not
an issue anymore; but part of the reason why we did stuff
like that in TAP.

Change-Id: Ib8f1989e398816ef071786240625fdeef908dce3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/185776
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>