MB-54516: Track the default collection's high legacy seqnos
Every collection tracks a high-seqno, and this value includes all
'events' DCP can send. E.g prepare, abort, mutation and system event.

Any 7.0 (collection aware) DCP client assert that their DCP stream will
reach this high-seqno. Collection enablement means that SeqnoAdvance
will skip any 'events' they don't opt-in to.

In the context of this change, a legacy DCP client is one which does
not enable collections, they cannot receive SeqnoAdvance, so KV
maintains two sequence numbers for the default collection which can be
exposed to such a client.

1) Max Visible Seqno (MVS)

Added in 7.1, this tracks the highest mutation a vbucket stores (and
which DCP would send). This does not increment for a prepare/abort
or system event.

In this commit - we now handle the system-event case as KV can now
generate a modify of the default collection - skipping an increment
of the default collection's MVS.

2) max legacy DCP seqno

New in this commit a new seqno which tracks the highest 'event' a
legacy DCP stream could theoretically reach. This will increment for
mutation, prepare and abort, but never for a system event.

Note: In this commit a warmup will lose this new seqno.

Change-Id: Id876e47099380dd90059d98534bc9e0b3b0c9b89
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/186296
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>