MB-55530 Expose compaction stats
Magma's compaction triggers have changed to use on disk uncompressed size for the data level. Knowing if the compactions have got relaxed/aggressive due to this is helpful. Hence expose compaction stats breakdown between keyIndex and seqIndex.

The fragmentation ratio also is internally tuned (static for now) to account for the checkpoint overheads. It will internally reduce the ratio for the seqIndex delta:data level. Hence expose the overhead as a stat.

Change-Id: I370347d48b32740fa9f19ab6d8e7b0a80f34d37f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/186789
Well-Formed: Restriction Checker
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>