MB-55467: Compaction passes HistoryMode to magma at WriteDocs()
The MagmaKVStore::compactDBInternal() might make a call into
magma::WriteDocs() by omitting the HistoryMode arg. By default the API
sets HisotryMode::Disabled, which has the effect of discarding the
history for the vbucket under compaction.

Fix that by passing the proper HistoryMode to the call.

Note: In the following we'll remove the default value for the
HistoryMode arg from the API for preventing any similar issue. Deferred
to a dedicated change.

Change-Id: If4d25d75852b0a7949ebbebcfb0bb0203fb6afd6
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/186506
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>