Refactor: Unify read/write callbacks for folly
Move the read/write callbacks used in the core to a
common location and use the same callbacks in the async client
used for testing.

After the move the code is isolated in such a way that it is
easy to write unit tests and possibly other benchmarks

Please note that the folly connections is not yet used in our
product (only a prototype which may be enabled by setting
environment variables) and in some of our unit tests.

Change-Id: I7ed3de87eac97b5b2830ca399b9e50c3257f9c0d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/186542
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>