MB-54516: Support warmup of the legacy max DCP seqno
This commit adds code to set the default collection's legacy max
seqno when a warmup occurs.

The technique used here is to attach to each generated modify event
(modify of the default collection) the current in-memory value of the
legacy max DCP seqno. Warmup can then look for this value if there is
a modify of the default collection stored in the vbucket KVStore.

The code utilises xattrs to store this extra metadata in only the
default collection modify.

Change-Id: Ie930ba65b77eed1ec0137cca9c08cdb93cb8e985
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/186556
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>